[DEBO04][DuongHT] update UI
