+++ src/pages/Admin/adminCreate/Branchs/BranchList/BranchList.tsx
@@ -1,0 +1,14 @@
+import { Box } from "@mui/material";
+import BranchHeader from "./BranchHeader";
+import BranchTable from "./BranchTable";
+
+function BranchList() {
+    return (
+        <Box>
+            <BranchHeader />
+            <BranchTable url={''} />
+        </Box>
+    );
+}
+
+export default BranchList;
