--- src/pages/Admin/adminCreate/CreateStaff/CreateStaff.tsx
@@ -1,10 +1,0 @@
-
-import StaffCreateBody from "./StaffCreateBody";
-
-function CreateStaff() {
-  return (
-    <StaffCreateBody></StaffCreateBody>
-  );
-}
-
-export default CreateStaff;
