+++ src/pages/Admin/adminCreate/Treatments/TreatmentList/TreatmentList.tsx
@@ -1,16 +1,16 @@
 import { Box } from "@mui/material";
 import TreatmentTable from "./TreatmentTable";
-import AdminCreateHeader from "../../components/AdminCreateHeader";
+import TreatmentHeaders from "./TreatmensHeader";
 import { API_ENDPOINTS } from "@/utils/api";
 
-function Treatments() {
+function TreatmentList() {
 
   return (
     <Box>
-      <AdminCreateHeader />
+      <TreatmentHeaders />
       <TreatmentTable url={API_ENDPOINTS.TREATMENT.GET_ALL_TREATMENT} />
     </Box>
   );
 }
 
-export default Treatments;
+export default TreatmentList;
